test: improve tests for `GradientBoosting` classifier & regressor
