+++ src/safeds/ml/classical/classification/__init__.py
@@ -3,7 +3,7 @@
 from ._ada_boost import AdaBoost
 from ._classifier import Classifier
 from ._decision_tree import DecisionTree
-from ._gradient_boosting_classification import GradientBoosting
+from ._gradient_boosting import GradientBoosting
 from ._k_nearest_neighbors import KNearestNeighbors
 from ._logistic_regression import LogisticRegression
 from ._random_forest import RandomForest
