+++ src/safeds/ml/classical/regression/__init__.py
@@ -3,7 +3,7 @@
 from ._ada_boost import AdaBoost
 from ._decision_tree import DecisionTree
 from ._elastic_net_regression import ElasticNetRegression
-from ._gradient_boosting_regression import GradientBoosting
+from ._gradient_boosting import GradientBoosting
 from ._k_nearest_neighbors import KNearestNeighbors
 from ._lasso_regression import LassoRegression
 from ._linear_regression import LinearRegression
